backend/dailyward/requirements.txt
django>=5.2.6
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
python-dotenv>=1.0.0
gunicorn>=20.1.0
psycopg2-binary>=2.9.6